houhaoyu@huawei.com
fix throw js error in har

+++ compiler/src/fast_build/ets_ui/rollup-plugin-ets-typescript.ts
@@ -98,7 +98,8 @@ export function etsTransform() {
     },
     moduleParsed(moduleInfo) {
       if (projectConfig.compileHar) {
-        if (moduleInfo.id && !moduleInfo.id.match(/node_modules/) && !moduleInfo.id.startsWith('\x00')) {
+        if (moduleInfo.id && !moduleInfo.id.match(new RegExp(projectConfig.packageDir)) &&
+          !moduleInfo.id.startsWith('\x00')) {
           const filePath: string = moduleInfo.id;
           const jsCacheFilePath: string = genTemporaryPath(filePath, projectConfig.moduleRootPath,
             process.env.cachePath, projectConfig);
@@ -117,8 +118,10 @@ export function etsTransform() {
     afterBuildEnd() {
       if (projectConfig.compileHar) {
         incrementalFileInHar.forEach((jsBuildFilePath, jsCacheFilePath) => {
-          const sourceCode: string = fs.readFileSync(jsCacheFilePath, 'utf-8');
-          writeFileSync(jsBuildFilePath, sourceCode);
+          if (fs.existsSync(jsCacheFilePath)) {
+            const sourceCode: string = fs.readFileSync(jsCacheFilePath, 'utf-8');
+            writeFileSync(jsBuildFilePath, sourceCode);
+          }
         });
       }
       shouldDisableCache = false;
